Disable automatic deduplication by default (#62)
As noted in the deprecation comment. Automatically
suffixing filenames with " (1)", " (2)" and so on is
a little difficult when dealing with directory nesting,
and represents a bit of "magic" that should, ideally, be
removed. However, some libraries might still be reliant
on it so we use the major version change to alter the default
value now.