Improve and update reading ZIPs over HTTP (#64)
Working with S3 it is better to determine the object size using a GET request for one byte.

The API surface of RemoteUncap is also made smaller. To test RemoteUncap we are much better off using an actual server instead of mocks - so that we are certain everything works. It only takes about half a second to spin up a Puma for that.

Reduce the number of Travis-tested Ruby versions, since what we care about is the lowest-supported and the latest-available (Ruby 2.7 PR will be a step in that direction too).