Raise on Streamer offset mismatch (#70)
To prevent a situation where a user might forget to use `sumulate_write`

This verification is performed when the central directory gets written -
we could fail earlier with it, but then we would need to walk all written
entries in order, after each added entry. This would make the algorithm
accidentally quadratic. If we want better performance we need to add an
additional offset tracking, and update it on every entry. This will allow
us to fail earlier, but is a bit of a drag to implement correctly.

To me it seems that failing at central directory output is a sensible compromise.