Reduce DeflatedWriter memory usage
By using the block syntax for deflate we generate smaller strings that
can more easily be garbage collected.

Due to a bug in JRuby 9.0 block syntax isn't used for finish, but that
doesn't really impact memory usage die to the low amount of data that is
usually flushed during finish.