Add write buffering in OutputEnumerator (#84)
The RackBody class spins up a Streamer which in turn spins up a ZipWriter, passing along an io-like object. The ZipWriter will << many extremely small strings into this io object, often only one or two bytes for local header values. Puma handles this rather poorly, since it essentially does body.each {|chunk| socket.syswrite(chunk)}. This leads to dozens/hundreds of very small write() syscalls and all the context switches to the kernel generates quite a bit of overhead.

We also optimize the write buffer a little bit so that it won't be buffering writes which are larger than the desired buffer size, but will pass them through untouched. We ensure the write buffer does not create copies of the string it hosts, to conserve memory, and we explain why we do this in the documentation.

On webserver use, the buffering OutputEnumerator will make the serving roughly 15 to 20 percent faster as opposed to not using buffering at all.

For all standard cases we are going to continue using unbuffered writes, as the performance actually worsens when buffering writes to the local filesystem (as it has some buffering on its own already).

Co-authored-by: Felix Bünemann <felix.buenemann@gmail.com>

Closes #78