Fix Rails streaming yet again
It turns out that there are a few things which were broken:

* The addition of Rack::ETag middleware would kill streaming responses
* The downstream HTTP proxy can request HTTP/1.0 and would then not support Transfer-Encoding: chunked

So: if the upstream proxy wants HTTP/1.0 we at least indicate to the user that there
is going to be buffering, so that they know where to look for configuration.

To bypass ETag and ContentLength middleware we add some special sauce. And this feature should
be tested against a real ActionController.