Add Tempfile output as well
If we buffer, it is actually better to buffer to a file so that we do not
have to generate the ZIP file twice _and_ use memory. While memory is cheap
these days, disk is still cheaper (and a better fit for temporary data which
won't be read all too quickly). And Rack can make use of X-Sendfile so that
the file will be delivered quicker to the client.