Autoload sub-modules
zip_tricks is a sizeable library, and some components are likely not to be used as often.
For example, someone willing to export ZIPs using zipline or RailsStreaming likely does
not need the FileReader or the RemoteUncap. That is code that has to be loaded, and it
will slow down application boot - for every application that loads zip_tricks. Why
not use something that has existed for ages and only load stuff we truly need?