Apply standardrb, mandate Ruby 2.6
standard is by now much more widespread than wetransfer_style.
The rubocop config of wetransfer_style was a consensus config and
has been bikeshedded over a lot, and changing it is unlikely in
the future. Using a standard (sic!) format will make the barrier to
contribution smaller. We also need to bump to 2.6 for running CI easier.

Installing older Rubies is hard, and this is, after all, a "new" gem.