Get rid of implicit chunked/Tempfile bodies (#6)
Following the discussion in https://github.com/julik/zip_kit/issues/2 -
it is probably better to let the server decide to chunk or not chunk the
response. On modern Rails, there is no Rack::ContentLength middleware in
place. It would be nice to detect it, but the setup would be too brittle.