Make RailsStreaming compatible with Live (#8)
ActionController::Live is a weird beast - but so be it. This makes
`zip_kit_stream` work with controllers which use Live too - which
requires a few tweaks.

This comes as a followup for https://github.com/julik/zip_kit/issues/7 -
it should be the responsibility of the library to ensure things still
work for users.